examples: fix time stamp prefix printf format string
The simples way to print time_t for 32bit and 64bit is to cast it to
unsigned long long. We could also use the time_t helpers for generating
a filename but that seems overkill for the example.

+++ examples/telemetry-listen.c
@@ -50,8 +50,8 @@ static void save_telemetry(nvme_ctrl_t c)
 		return;
 
 	s = time(NULL);
-	ret = snprintf(buf, sizeof(buf), "/var/log/%s-telemetry-%ld",
-		nvme_ctrl_get_subsysnqn(c), s);
+	ret = snprintf(buf, sizeof(buf), "/var/log/%s-telemetry-%llu",
+		       nvme_ctrl_get_subsysnqn(c), (unsigned long long)s);
 	if (ret < 0) {
 		free(log);
 		return;
